requirements.txt
Django==5.1.4 
django-tailwind==3.8.0
django-browser-reload==1.17.0
django-filter==24.3
djangorestframework==3.15.2
markdown==3.7
django-debug-toolbar==4.4.6
django-tinymce==4.1.0